+++ changelog.md
@@ -1,8 +1,17 @@
-# 1.0.0
-- Release a mod
+# 1.1.0
+- Added an in-game demonlist to the level search layer.
+
+# 1.0.4
+- Just a little patch
+
+# 1.0.3
+- Fixed mod desc (demoblist => demonlist)
+
+# 1.0.2
+- Fixed about.md
 
 # 1.0.1
 - Fixed interface (if the level position isn't found, the mod's UI elements will be hidden.)
 
-# 1.0.2
-- Fixed about.md
+# 1.0.0
+- Release a mod